microservices/auth/requirements.txt
protobuf==5.29.4
grpcio==1.71.0
googleapis-common-protos==1.69.2
PyJWT==2.10.1
psycopg2-binary